% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab06.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solve the knapsack problem using a brute force approach.}
\usage{
brute_force_knapsack(x, W, fast = FALSE)
}
\arguments{
\item{x}{A data frame object, containing two variables \code{v} and \code{w}. \code{v} contains the value of the object n and \code{w} its value.}

\item{W}{An integer, representing the capacity of the knapsack.}
}
\value{
A list with two objects: \code{$value} which is the value of the knapsack and an object \code{$elements} which are the elements contained in the knapsack.
}
\description{
A function that solves the knapsack problem using a brute force approach.
}
\details{
none
}
\examples{
\\code{set.seed(42)}
\\code{x <- data.frame(w=sample(1:4000, size = 10, replace = TRUE), v=runif(n = 10, 0, 10000))}
\\code{W <- 3500}
\\code{brute_force_knapsack(x, W)}
}
