% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab06.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{using some heuristic or common sense knowledge to generate a sequence of suboptimum that hopefully converges to an optimum value.}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe with two columns: the values (v) and the weights (w) of each item to put in the knapsack.}

\item{W}{A positive number representing the knapsack size..}
}
\value{
A list of two elements: a positive number with the maximum knapsack \code{value} and a vector of all the \code{elements} in the knapsack size.
}
\description{
A function that solves the knapsack problem using a greedy approach.
}
\details{
none
}
\examples{
set.seed(42)
x <- data.frame(w=sample(1:4000, size = 800, replace = TRUE), v=runif(n = 10, 0, 10000))
W <- 3500
greedy_knapsack(x, W)
}
