% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab06.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming in in knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A dataframe with two columns: the values (v) and the weights (w) of each item to put in the knapsack.}

\item{W}{A positive number representing the knapsack size.}
}
\value{
A list of two elements: a positive number with the maximum knapsack \code{value} and a vector of all the \code{elements} in the knapsack size.
}
\description{
A function that solves the knapsack problem using a dynamic approach.
}
\details{
none
}
