% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team05Lab06.R
\docType{package}
\name{Team05Lab06}
\alias{Team05Lab06}
\alias{Team05Lab06-package}
\title{A package for solving 0/1 Knapsack problem.}
\description{
This package contains three functions for solving 0/1 Knapsack problem using a different approach in each function. The functions are \code{brute_force_knapsack()}, \code{knapsack_dynamic()} and \code{greedy_knapsack()}.
}
\details{
The knapsack problem is a problem of optimization. Given a set of items, each item with a weight and a value, the objective is to fill a knapsack with the subset that offers the best value without exceeding the knapsack weight capacity. More details are given in \href{https://en.wikipedia.org/wiki/Knapsack_problem}{Wikipedia.}
}
\author{
\strong{Maintainer}: Jose Mendez \email{josme478@student.liu.se}

Authors:
\itemize{
  \item Sara Salahshour \email{sarsa293@student.liu.se}
  \item Atieh Khaleghi \email{atikh281@student.liu.se}
}

}
